requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
certifi==2022.9.24       
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dominate==2.7.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
flask-restx==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==2.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.0
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
pyodbc==4.0.34
pyparsing==3.0.9
pyrsistent==0.19.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.12
visitor==0.1.3
Werkzeug==2.0.3
pytest==6.2.5